ffi: Move auto-generated bindings to an inner module
This commit is a preparation to a bigger change to wrapper few `ffi`
calls.

We are adding the generated bindings inside an inner `ffi::generated`
module and re-exporting it as `ffi` so requiring no code changes to use
it.

This also change the symbols visibility to be pub(crate) as we do not
intent people to call the `ffi` module directly.

Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>