Improve Windows GitHub Action
Add an additional matrix parameter named 'linkage'. This parameter
has values corresponding to the vcpkg triplet used to install
libarchive, but effectively represents whether a static or dynamic
build is being run/created.
This does mean that the Windows action now creates 6 jobs instead of 3,
a 100% increase, though they run in parallel instead.
Depending on which value of the 'linkage' parameter the job is run
with, different environment variables are emitted to cargo, thus
changing how the library is built as well as the version of libarchive
installed by vcpkg.