Fix absolute paths extracting outside of destination directory by stripping leading '/'
Previously, an absolute entry path would be joined with the destination, but in Rust joining an absolute path onto another one will just return the absolute path which results in files being extracted outside of the destination directory. This commit changes the behaviour to strip the leading '/' from an absolute path, matching behaviour from bsdtar and gnu tar which will extract the files into the destination directory.